+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Changes in version 3.4**
 
+- Changed PRCDA and UIHO algorithms to use County at DX Analysis instead of reported County at DX.
 - Updated NAACCR XML library from version 1.4.10 to version 1.4.17
 - Updated CSV library from version 5.2 to version 5.4.
 - Updated Apache commons-lang library from version 3.9 to version 3.11.